fix: moved pages from components to pages folder
