+++ src/pages/Loading.jsx
@@ -1,0 +1,13 @@
+import LoadingSpinner from "../components/LoadingSpinner"
+
+import '../css/loading.css'
+
+const Loading = () => {
+    return (
+        <div id="spinner-container">
+            <LoadingSpinner />
+        </div>
+    )
+} 
+
+export default Loading
